Update: Added Section 2.3

+++ Unit 2/Section 2.2/ClassesV2.java
@@ -1,7 +1,7 @@
 public class ClassesV2 {
     @SuppressWarnings("unused")
     public static void main(String[] args) {
-        Employee BadBoyRajat = new Employee("Rajat", "1298hasd", 1.0);
+        Employee BadBoyRajat = new Employee("Rajat", "1298hasd", (double) 1);
 
     }
 }
